feat(vue): Refactored mixins. Added fabAPI mixin. Supported onClick props.
