+++ bindings/vue/src/components/index.js
@@ -8,24 +8,24 @@ export { default as VOnsSplitter } from './VOnsSplitter.vue';
 export { default as VOnsSplitterSide } from './VOnsSplitterSide.vue';
 export { default as VOnsSplitterContent } from './VOnsSplitterContent.vue';
 export { default as VOnsRange } from './VOnsRange.vue';
+export { default as VOnsSpeedDial } from './VOnsSpeedDial.vue';
 
 // Generic components
 import VGeneric from './VGeneric.vue';
+import { dialogAPI, fabAPI } from '../internal/mixins/api';
+import { clickable, hasOptions } from '../internal/mixins/common';
 
 const extend = (component, mixins = []) => ({ name: 'v-ons-' + component, mixins, extends: VGeneric });
 
 export const VOnsPage = extend('page');
 export const VOnsToolbar = extend('toolbar');
 export const VOnsToolbarButton = extend('toolbar-button');
-export const VOnsBackButton = extend('back-button');
 export const VOnsButton = extend('button');
 export const VOnsCarousel = extend('carousel');
 export const VOnsCarouselItem = extend('carousel-item');
 export const VOnsIcon = extend('icon');
 export const VOnsSwitch = extend('switch');
 export const VOnsBottomToolbar = extend('bottom-toolbar');
-export const VOnsFab = extend('fab');
-export const VOnsSpeedDial = extend('speed-dial');
 export const VOnsSpeedDialItem = extend('speed-dial-item');
 export const VOnsList = extend('speed-list');
 export const VOnsListItem = extend('speed-list-item');
@@ -36,11 +36,8 @@ export const VOnsCol = extend('col');
 export const VOnsProgressBar = extend('progress-bar');
 export const VOnsSplitterMask = extend('splitter-mask');
 export const VOnsPullHook = extend('pull-hook');
-
-
-// Generic Dialogs
-import { dialogsAPI } from '../internal/mixins/dialogs';
-
-export const VOnsDialog = extend('dialog', [dialogsAPI]);
-export const VOnsModal = extend('modal', [dialogsAPI]);
+export const VOnsFab = extend('fab', [fabAPI]);
+export const VOnsBackButton = extend('back-button', [clickable, hasOptions]);
+export const VOnsDialog = extend('dialog', [dialogAPI]);
+export const VOnsModal = extend('modal', [dialogAPI]);
