+++ bindings/vue/src/internal/mixins/api.js
@@ -1,0 +1,18 @@
+import { visibilityToggle, clickable, hasOptions } from './common';
+
+const dialogAPI = {
+  mixins: [visibilityToggle, hasOptions],
+
+  mounted() {
+    this.$el._cancel = () => {
+      this.$emit('mask');
+    };
+  }
+};
+
+const fabAPI = {
+  mixins: [clickable, visibilityToggle]
+};
+
+export { dialogAPI, fabAPI };
+
